ADDED : Test Case To Find Top Bowler Name By Average

+++ src/main/java/com/bridgelabz/test/IPLWicketsCSV.java
@@ -34,7 +34,7 @@ public class IPLWicketsCSV {
     public int fiveWickets;
 
     @CsvBindByName(column = "Econ", required = true)
-    public int economy;
+    public double economy;
 
 
     @Override
